* move shaders into own files * move view rect calculation out of vertex shader * do more intricate aspect ratio correction
In the end, it is not feasible to keep a proper pixel ratio with arbitrary resolutions just by correcting the aspect ratio.
The mode 13h resolution 320x200 is 16:10 and so fitting it in 1280x720 will mean squeezing the pixels so they are less than
exactly 4 times larger making it no longer one-to-one. This can be fixed by:

1. Using the pixel filtering from the Handmade Chat to sample the pixels more one-to-one
2. Not using 320x200 as the internal resolution all the time. Instead work backwards and establish
   a resolution based on the screen resolution such that the pixels are always scaled by integer.
   This means a more dynamic UI/game
3. Cropping instead of letterboxing

During testing I will just use 1280x800.